+++ packages/plugin/src/Library/Composer/Components/AbstractField.php
@@ -756,7 +756,7 @@ abstract class AbstractField implements FieldInterface, \JsonSerializable
             }
         }
 
-        if (!empty($value) && !$this->isHidden()) {
+        if ('' !== $value && !$this->isHidden()) {
             static $validator;
 
             if (null === $validator) {